refine evaluation rendering scripts

+++ render_360.py
@@ -1,38 +1,30 @@
-#
-# Copyright (C) 2023, Inria
-# GRAPHDECO research group, https://team.inria.fr/graphdeco
-# All rights reserved.
-#
-# This software is free for non-commercial, research and evaluation use 
-# under the terms of the LICENSE.md file.
-#
-# For inquiries contact  george.drettakis@inria.fr
-#
 
-import torch
-import subprocess
-import shutil
-import math
-from scene import Scene
-import os
 import gc
-from tqdm import tqdm
-from os import makedirs
-from gaussian_renderer import render
-import torchvision
-from utils.general_utils import safe_state
+import math
+import os
+import shutil
+import subprocess
 from argparse import ArgumentParser
-from arguments import ModelParams, PipelineParams, OptimizationParams, get_combined_args
-from gaussian_renderer import GaussianModel
-from gaussian_renderer import flashsplat_render
-from utils.wheatgs_helper import render_360, multi_instance_opt
+from os import makedirs
 
 import numpy as np
 import torch
-from utils.graphics_utils import getWorld2View2, getProjectionMatrix
+import torchvision
+from tqdm import tqdm
+
+from arguments import ModelParams, OptimizationParams, PipelineParams, get_combined_args
+from gaussian_renderer import GaussianModel, flashsplat_render, render
+from scene import Scene
 from scene.cameras import MiniCam
+from utils.general_utils import safe_state
+from utils.graphics_utils import getProjectionMatrix, getWorld2View2
+from utils.wheatgs_helper import multi_instance_opt, render_360
 
 #### Begin of 360-degree camera trajectory copied from gsgen ####
+# These two functions are adapted from the implementation in:
+# "GSGEN: Text-to-3D using Gaussian Splatting"
+# Original Code: https://github.com/gsgen3d/gsgen
+
 def get_c2w_from_up_and_look_at(up, look_at, pos):
     up = up / np.linalg.norm(up)
     z = look_at - pos
@@ -76,7 +68,7 @@ def opt_w_masks(viewpoint_cam, gaussians, pipe, background, obj_masks, obj_num=N
     used_count = render_pkg["used_count"].detach().cpu()
     return used_count, obj_num
 
-def render_all(dataset : ModelParams, pipeline : PipelineParams, exp_name):
+def render_wheat_field(dataset : ModelParams, pipeline : PipelineParams, exp_name, n_frames=100, framerate=10, save_frames=False):
     with torch.no_grad():
         gaussians = GaussianModel(dataset.sh_degree)
         scene = Scene(dataset, gaussians, load_iteration=None, shuffle=False)
@@ -115,7 +107,7 @@ def render_all(dataset : ModelParams, pipeline : PipelineParams, exp_name):
             render_path = os.path.join(os.path.dirname(wheat_head_dir), "3DWheat", ply_id)
             makedirs(render_path, exist_ok=True)
 
-            c2ws = get_camera_path_fixed_elevation(n_frames=100, n_circles=1, camera_distance=camera_distance, cam_center=gs_centroid, elevation=15)
+            c2ws = get_camera_path_fixed_elevation(n_frames=n_frames, n_circles=1, camera_distance=camera_distance, cam_center=gs_centroid, elevation=15)
             for idx, c2w in enumerate(c2ws):
                 c2w = np.vstack([c2w, [0.0, 0.0, 0.0, 1.0]])
                 w2c = np.linalg.inv(np.float64(c2w))
@@ -140,68 +132,12 @@ def render_all(dataset : ModelParams, pipeline : PipelineParams, exp_name):
                 "-pix_fmt", "yuv420p",
                 output_video
             ], cwd=render_path)
-            # shutil.rmtree(render_path)
-            
-def render_sets(dataset : ModelParams, iteration : int, pipeline : PipelineParams, which_wheat_head=None):
-    with torch.no_grad():
-        gaussians = GaussianModel(dataset.sh_degree)
-        scene = Scene(dataset, gaussians, load_iteration=iteration, shuffle=False, which_wheat_head=which_wheat_head)
-        bg_color = [1,1,1] if dataset.white_background else [0, 0, 0]
-        background = torch.tensor(bg_color, dtype=torch.float32, device="cuda")
+            if not save_frames:
+                shutil.rmtree(render_path)
 
-        og_views = scene.getTrainCameras()
-        og_view = og_views[0]
-        
-        if which_wheat_head is not None:
-            width, height = math.floor(og_view.image_width / 3), math.floor(og_view.image_height / 3)
-            fovy, fovx = og_view.FoVy / 7, og_view.FoVx / 7
-        else:
-            width, height = math.floor(og_view.image_width / 2), math.floor(og_view.image_height / 2)
-            fovy, fovx = og_view.FoVy, og_view.FoVx
-        znear, zfar = og_view.znear, og_view.zfar
-        print(f"Fixed parameters width: {width} height {height} fovy {fovy} fovx {fovx}")
-
-        gs_centroid = torch.mean(gaussians.get_xyz, dim=0).cpu().tolist()
-        scene_radius = scene.cameras_extent
-        print(f"Gaussians centroid {gs_centroid}, Scene radius {scene_radius}")
-
-        if which_wheat_head is None:
-            camera_distance = scene_radius*4
-            render_path = os.path.join(dataset.model_path, "render360", "wheat-field")
-        else:
-            camera_distance = scene_radius * 0.75
-            render_path = os.path.join(dataset.model_path, "render360", str(int(which_wheat_head)).zfill(4))
-        makedirs(render_path, exist_ok=True)
-        c2ws = get_camera_path_fixed_elevation(n_frames=200, n_circles=1, camera_distance=camera_distance, cam_center=gs_centroid, elevation=45)
-        for idx, c2w in enumerate(tqdm(c2ws, desc="Rendering progress")):
-            c2w = np.vstack([c2w, [0.0, 0.0, 0.0, 1.0]])
-            w2c = np.linalg.inv(np.float64(c2w))
-            world_view_transform = torch.tensor(w2c.astype(np.float32)).transpose(0, 1).cuda()
-            projection_matrix = getProjectionMatrix(znear, zfar, fovx, fovy).transpose(0,1).cuda()
-            full_proj_transform = (world_view_transform.unsqueeze(0).bmm(projection_matrix.unsqueeze(0))).squeeze(0)
-            view = MiniCam(width, height, fovy, fovx, znear, zfar, world_view_transform, full_proj_transform)
-            render_pkg = render(view, gaussians, pipeline, background)
-            rendering = render_pkg["render"]
-            torchvision.utils.save_image(rendering, os.path.join(render_path, '{0:05d}'.format(idx) + ".png"))
-        if which_wheat_head is not None:
-            output_video = os.path.join(os.path.dirname(render_path), f"{int(which_wheat_head):04}.mp4")
-        else:
-            output_video = os.path.join(os.path.dirname(render_path), f"wheat-field-d5-e45.mp4")
-        subprocess.run([
-            "ffmpeg",
-            "-loglevel", "error",
-            "-start_number", "0",
-            "-i", "%05d.png",  
-            "-vf", "scale=iw-mod(iw\,2):ih-mod(ih\,2),fps=5",
-            "-vcodec", "libx264",
-            "-pix_fmt", "yuv420p",
-            output_video
-        ], cwd=render_path)
-        shutil.rmtree(render_path)
-
-def render_3d_seg(dataset : ModelParams, pipeline : PipelineParams, opt, load_iteration=-1):
-    seg2d_dir = os.path.join(dataset.model_path, "wheat-head", "run_test", "2DSeg")
-    out_dir = os.path.join(dataset.model_path, "wheat-head", "run_test", "3DSeg")
+def render_wheat_head(dataset : ModelParams, pipeline : PipelineParams, exp_name, n_frames=100, framerate=10, load_iteration=-1, save_frames=False):
+    seg2d_dir = os.path.join(dataset.model_path, "wheat-head", exp_name, "2DSeg")
+    out_dir = os.path.join(dataset.model_path, "wheat-head", exp_name, "3DSeg")
     os.makedirs(out_dir, exist_ok=True)
     gaussians = GaussianModel(dataset.sh_degree)
     try:
@@ -210,7 +146,7 @@ def render_3d_seg(dataset : ModelParams, pipeline : PipelineParams, opt, load_it
         pass
     print(f"Load iteration {load_iteration}, Resolution {dataset.resolution}")
     scene = Scene(dataset, gaussians, load_iteration=load_iteration, shuffle=False)
-    gaussians.training_setup(opt)
+    # gaussians.training_setup(opt)
     print(f"Loaded point cloud size: {len(gaussians.get_xyz)}")
     bg_color = [1, 1, 1] if dataset.white_background else [0, 0, 0]
     background = torch.tensor(bg_color, dtype=torch.float32, device="cuda")
@@ -239,12 +175,12 @@ def render_3d_seg(dataset : ModelParams, pipeline : PipelineParams, opt, load_it
         torch.cuda.empty_cache()
     print(f"All counts: {all_counts.shape}")
     slackness = 0.0
-    torch.save(all_counts.detach().cpu(), os.path.join(dataset.model_path, "wheat-head", "run_test", "all_counts.pth"))
+    torch.save(all_counts.detach().cpu(), os.path.join(dataset.model_path, "wheat-head", exp_name, "all_counts.pth"))
     all_obj_labels = multi_instance_opt(all_counts, slackness)
     print("All_obj_labels: ", all_obj_labels.shape)
     # Save for future rendering
-    torch.save(all_obj_labels.detach().cpu(), os.path.join(dataset.model_path, "wheat-head", "run_test", "all_obj_labels.pth"))
-    output_video = render_360(viewpoint_stack[0], scene.cameras_extent, out_dir, 200, 10, gaussians, pipeline, background, all_obj_labels=all_obj_labels)
+    torch.save(all_obj_labels.detach().cpu(), os.path.join(dataset.model_path, "wheat-head", exp_name, "all_obj_labels.pth"))
+    output_video = render_360(viewpoint_stack[0], scene.cameras_extent, out_dir, n_frames, framerate, gaussians, pipeline, background, all_obj_labels=all_obj_labels)
 
 if __name__ == "__main__":
     # Set up command line argument parser
@@ -252,17 +188,23 @@ if __name__ == "__main__":
     model = ModelParams(parser)
     pipeline = PipelineParams(parser)
     opt = OptimizationParams(parser)
-    parser.add_argument("--iteration", default=-1, type=int)
-    parser.add_argument("--which_wheat_head", type=str, default=None)
-    parser.add_argument("--exp_name", type=str, default=None)
+    parser.add_argument("--iteration", default=-1, type=int, help="iteration of OG 3DGS to load")
+    parser.add_argument("--render_type", type=str, default=None, help="render type: field (whole wheat field) or head (all individual wheat heads)")
+    parser.add_argument("--exp_name", type=str, default=None, help="experiment name of 3D segmentation to load")
+    parser.add_argument("--n_frames", type=int, default=None, help="number of frames to render")
+    parser.add_argument("--framerate", type=int, default=None, help="framerate of the rendered video")
+    parser.add_argument("--save_frames", action="store_true", help="If specified, save frames in addition to output video")
     args = get_combined_args(parser)
-    print(f"Rendering {args.model_path} for wheat head {args.which_wheat_head}")
-    if args.which_wheat_head == "all":
-        print("Render All Wheat Heads")
-        render_all(model.extract(args), pipeline.extract(args), args.exp_name)
-    elif args.which_wheat_head == "3dseg":
-        print("Render 3D Segmentation")
-        render_3d_seg(model.extract(args), pipeline.extract(args), opt.extract(args))
+    print(f"Rendering {args.model_path} for 3D segmentation experiment {args.exp_name}, Option: {args.which_wheat_head}")
+    if args.n_frames is None:
+        args.n_frames = 100
+    if args.framerate is None:
+        args.framerate = 10
+    if args.render_type == "field":
+        print("Render the 3D segmentation on the whole wheat field")
+        render_wheat_field(model.extract(args), pipeline.extract(args), args.exp_name, args.n_frames, args.framerate, args.save_frames)
+    elif args.render_type == "head":
+        print("Render each individual segmented wheat head")
+        render_wheat_head(model.extract(args), pipeline.extract(args), args.exp_name, args.n_frames, args.framerate, args.save_frames)
     else:
-        pass
-        # render_sets(model.extract(args), args.iteration, pipeline.extract(args), args.which_wheat_head)
+        raise ValueError(f"Invalid render type: either field or head")
